af: Gate native Swift/Objective-C code per platform [SDK-4700] 1/3 (#346)
