+++ auth0_flutter/darwin/Classes/Auth0FlutterPlugin.h
@@ -1,0 +1,7 @@
+#if TARGET_OS_IOS
+    #import <Flutter/Flutter.h>
+#else
+    #import <FlutterMacOS/FlutterMacOS.h>
+#endif
+@interface Auth0FlutterPlugin : NSObject<FlutterPlugin>
+@end
