+++ auth0_flutter/darwin/Classes/Auth0FlutterPlugin.m
