+++ auth0_flutter/darwin/Classes/AuthAPI/AuthAPIExtensions.swift
@@ -1,5 +1,10 @@
-import Flutter
 import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
 
 enum AuthAPIErrorFlag: String, CaseIterable {
     case isMultifactorRequired
