+++ auth0_flutter/darwin/Classes/AuthAPI/AuthAPIResetPasswordMethodHandler.swift
@@ -1,0 +1,39 @@
+import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
+
+struct AuthAPIResetPasswordMethodHandler: MethodHandler {
+    enum Argument: String {
+        case email
+        case connection
+        case parameters
+    }
+
+    let client: Authentication
+
+    func handle(with arguments: [String: Any], callback: @escaping FlutterResult) {
+        guard let email = arguments[Argument.email] as? String else {
+            return callback(FlutterError(from: .requiredArgumentMissing(Argument.email.rawValue)))
+        }
+        guard let connection = arguments[Argument.connection] as? String else {
+            return callback(FlutterError(from: .requiredArgumentMissing(Argument.connection.rawValue)))
+        }
+        guard let parameters = arguments[Argument.parameters] as? [String: Any] else {
+            return callback(FlutterError(from: .requiredArgumentMissing(Argument.parameters.rawValue)))
+        }
+
+        client
+            .resetPassword(email: email, connection: connection)
+            .parameters(parameters)
+            .start {
+                switch $0 {
+                case .success: callback(nil)
+                case let .failure(error): callback(FlutterError(from: error))
+                }
+            }
+    }
+}
