+++ auth0_flutter/darwin/Classes/CredentialsManager/CredentialsManagerExtensions.swift
@@ -1,5 +1,10 @@
-import Flutter
 import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
 
 extension FlutterError {
     convenience init(from credentialsManagerError: CredentialsManagerError) {
