+++ auth0_flutter/darwin/Classes/CredentialsManager/CredentialsManagerGetMethodHandler.swift
@@ -1,5 +1,10 @@
-import Flutter
 import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
 
 struct CredentialsManagerGetMethodHandler: MethodHandler {
     enum Argument: String {
