+++ auth0_flutter/darwin/Classes/CredentialsManager/CredentialsManagerHandler.swift
@@ -1,5 +1,10 @@
-import Flutter
 import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
 
 // MARK: - Providers
 
@@ -22,8 +27,15 @@ public class CredentialsManagerHandler: NSObject, FlutterPlugin {
 
     public static func register(with registrar: FlutterPluginRegistrar) {
         let handler = CredentialsManagerHandler()
+
+        #if os(iOS)
         let channel = FlutterMethodChannel(name: CredentialsManagerHandler.channelName,
                                            binaryMessenger: registrar.messenger())
+        #else
+        let channel = FlutterMethodChannel(name: CredentialsManagerHandler.channelName,
+                                           binaryMessenger: registrar.messenger)
+        #endif
+
         registrar.addMethodCallDelegate(handler, channel: channel)
     }
