+++ auth0_flutter/darwin/Classes/CredentialsManager/CredentialsManagerModels.swift
