+++ auth0_flutter/darwin/Classes/HandlerError.swift
