+++ auth0_flutter/darwin/Classes/Properties.swift
