+++ auth0_flutter/darwin/Classes/WebAuth/WebAuthHandler.swift
@@ -1,5 +1,10 @@
-import Flutter
 import Auth0
+
+#if os(iOS)
+import Flutter
+#else
+import FlutterMacOS
+#endif
 
 // MARK: - Providers
 
@@ -18,8 +23,15 @@ public class WebAuthHandler: NSObject, FlutterPlugin {
 
     public static func register(with registrar: FlutterPluginRegistrar) {
         let handler = WebAuthHandler()
+
+        #if os(iOS)
         let channel = FlutterMethodChannel(name: WebAuthHandler.channelName,
                                            binaryMessenger: registrar.messenger())
+        #else
+        let channel = FlutterMethodChannel(name: WebAuthHandler.channelName,
+                                           binaryMessenger: registrar.messenger)
+        #endif
+
         registrar.addMethodCallDelegate(handler, channel: channel)
     }
