+++ auth0_flutter/darwin/Classes/WebAuth/WebAuthModels.swift
@@ -1,12 +1,13 @@
+#if os(iOS)
 enum SafariViewControllerProperty: String, CaseIterable {
     case presentationStyle
 }
 
 struct SafariViewController {
     var presentationStyle: UIModalPresentationStyle = UIModalPresentationStyle.fullScreen
-    
+
     static let key = "safariViewController";
-    
+
     init(from dictionary: [String: Any?]) {
         if let presentationStyle = dictionary[SafariViewControllerProperty.presentationStyle] as? Int,
             let uiModalPresentationStyle = UIModalPresentationStyle(rawValue: presentationStyle) {
@@ -14,3 +15,4 @@ struct SafariViewController {
         }
     }
 }
+#endif
