+++ auth0_flutter/example/ios/Podfile
@@ -28,7 +28,7 @@ require File.expand_path(File.join('packages', 'flutter_tools', 'bin', 'podhelpe
 flutter_ios_podfile_setup
 
 target 'Runner' do
-  platform :ios, '12.0'
+  platform :ios, '13.0'
 
   use_frameworks!
   use_modular_headers!
