--- auth0_flutter/ios/Classes/Auth0FlutterPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface Auth0FlutterPlugin : NSObject<FlutterPlugin>
-@end
